audio-kernel: Fix compilation time issue
In incremental (dirty) kernel builds, compiling audio drivers takes
significantly longer or may even freeze. The export statement causes
unnecessary exports, which leads to this issue. Removing and adding
export statements to the conf files for a specific platform, as it was
done for Kona in the commit 4025a25 ("makefile: kona: remove make export
<all variables> instances") should fix this issue.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>
Change-Id: I4a52febc7852342946832b01638db023c3602711